+++ src/conversions.jl
@@ -195,7 +195,7 @@ function convert_arguments(::Type{<: Lines}, rect::Rect3{T}) where {T}
     PT = Point3{float_type(T)}
     points = decompose(PT, rect)
     push!(points, PT(NaN)) # use to separate linesegments
-    return (points[[1, 2, 3, 4, 1, 5, 6, 2, 9, 6, 8, 3, 9, 5, 7, 4, 9, 7, 8]],)
+    return (points[[1, 3, 4, 2, 1, 5, 6, 2, 9, 6, 8, 7, 5, 9, 8, 4, 9, 7, 3]],)
 end
 """
 
@@ -2003,6 +2003,8 @@ convert_attribute(value, ::key"diffuse") = Vec3f(value)
 convert_attribute(value, ::key"specular") = Vec3f(value)
 
 convert_attribute(value, ::key"backlight") = Float32(value)
+
+convert_attribute(value, ::key"depth_shift") = Float32(value)
 
 
 # SAMPLER overloads